src/cloud/requirements.txt
nats-py
boto3